undefined 체크 추가

+++ server/routes/template.js
@@ -136,18 +136,23 @@ router.post('/deleteTemplate', (req, res) => {
           console.log('File has been Deleted');    
           
           // 썸네일 이미지 삭제 
-          fs.access(template.thumbnail, fs.constants.F_OK, (err) => { // A
-            if (err) return console.log('삭제할 수 없는 파일입니다');
-            fs.unlink(template.thumbnail, (err) => err ?
-              console.log(err) : console.log(`${template.thumbnail} 를 정상적으로 삭제했습니다`));
-          });
+          if (template.thumbnail) {
+            fs.access(template.thumbnail, fs.constants.F_OK, (err) => { // A
+              if (err) return console.log('삭제할 수 없는 파일입니다');
+              fs.unlink(template.thumbnail, (err) => err ?
+                console.log(err) : console.log(`${template.thumbnail} 를 정상적으로 삭제했습니다`));
+            });
+          }
 
           // 사용자 템플릿 삭제
-          fs.access(template.customRef, fs.constants.F_OK, (err) => { // A
-            if (err) return console.log('삭제할 수 없는 파일입니다');
-            fs.unlink(template.customRef, (err) => err ?
-              console.log(err) : console.log(`${template.customRef} 를 정상적으로 삭제했습니다`));
-          });
+          if (template.customRef) {
+            fs.access(template.customRef, fs.constants.F_OK, (err) => { // A
+              if (err) return console.log('삭제할 수 없는 파일입니다');
+              fs.unlink(template.customRef, (err) => err ?
+                console.log(err) : console.log(`${template.customRef} 를 정상적으로 삭제했습니다`));
+            });
+          }
+
         });
       });
